[MIN-153] Remove setCollateralFactor.
